Refactor evaluation test suite to run on single file sources
