+++ tests/jballerina-debugger-integration-test/src/test/resources/testng.xml
@@ -57,8 +57,10 @@ under the License.
             <class name="org.ballerinalang.debugger.test.adapter.variables.VariableQueryTest"/>
 
             <!--Debugger Expression Evaluation Tests-->
-            <class name="org.ballerinalang.debugger.test.adapter.evaluation.ExpressionEvaluationTest"/>
-            <class name="org.ballerinalang.debugger.test.adapter.evaluation.ExpressionEvaluationNegativeTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.evaluation.PackageEvaluationTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.evaluation.PackageEvaluationNegativeTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.evaluation.SingleFileEvaluationTest"/>
+            <class name="org.ballerinalang.debugger.test.adapter.evaluation.SingleFileEvaluationNegativeTest"/>
             <class name="org.ballerinalang.debugger.test.adapter.evaluation.DependencyEvaluationTest"/>
 
             <!--Debug Completions Tests-->
